Fixes the spritemap arrows
The arrows which appear next to various headers to indicate the ability
to expand sections of the site have been broken for some time. This was
due to styles cascading into existing styles and changing a variety of
spacing, font-size, and line-height declarations.

- Add specific CSS to the various H5 rules to restore the original
style of the elements, fixing the spritemap arrows issue.