Fixes 500 on invalid post to DataStorageView
Currently POSTs to the DataStorageView (that saves progress to a
worksheet) that contain an invalid ID will cause a 500 error, e.g.

compare-financial-aid-and-college-costs/api/worksheet/invalid.json

This produces the following error:

DoesNotExist: Worksheet matching query does not exist

This PR handles the case of an invalid worksheet ID and instead returns
a 404 to the frontend. This should never happen during normal usage of
the tool, but prevents 500s due to crawlers or maliciously crafted URLs.