Restore storage of worksheet data
After we upgraded to Django 2.2 on May 18, the paying-for-college comparisontool
stopped saving worksheet json data properly. Because of the way Django
handles request body encoding, we need to decode content from a POST request
before saving it to the database.

Testing

1. Visit the comparison page at http://localhost:8000/paying-for-college/compare-financial-aid-and-college-cost/,
click "Get started" and follow the steps for picking up to 3 schools to compare.
To be able to pick schools, you'll need to have Elasticsearch running with a populated index.
If you haven't indexed the comparisontool data locally, you can do that with a manage.py command:

```bash
./cfgov/manage.py update_index comparisontool
```

2. After choosing your school(s), click "Continue" near the top right of the page.

3. Click the "Save your work" button

You should get an email dialog. Rather than attempting to email it (which will print it
to console in your runserver tab), copy the URL field from the email dialog.

Open a new browser tab, paste in the URL, and remove the "s" from "https" prefix.

Your request should return the restored worksheet, which proves the json data
made the round trip from your initial browser tab to the cfgov database and back.

Currently, if you try this routine on www.consumerfinance.gov,
you will encounter a server error when saving the worksheet.

Tox note: Since we've moved to Django 2, I removed Django 1.11 from the tox matrix.